pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypted-mail"
version = "0.1.0"
description = "Windows-first desktop app for encrypting text and sending it via Gmail."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-api-python-client>=2.169.0",
  "google-auth>=2.39.0",
  "google-auth-oauthlib>=1.2.2",
  "keyring>=25.6.0",
  "PyNaCl>=1.5.0",
  "PySide6>=6.9.0",
]

[project.optional-dependencies]
dev = [
  "Pillow>=11.2.1",
  "pyinstaller>=6.15.0",
  "pytest>=8.3.5",
  "pytest-qt>=4.4.0",
]

[project.scripts]
crypted-mail = "crypted_mail.desktop.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"crypted_mail.assets" = ["*.ico", "*.svg"]

[tool.pytest.ini_options]
testpaths = ["tests"]
